riskcast-v16-main/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
anthropic>=0.25.0
httpx>=0.25.0
requests>=2.31.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
numpy>=1.24.0
scipy>=1.10.0
jinja2>=3.1.0
itsdangerous>=2.1.0
# MySQL Database
sqlalchemy>=2.0.0
pymysql>=1.1.0
cryptography>=41.0.0
